+++ panther/__init__.py
@@ -1,6 +1,6 @@
 from panther.main import Panther  # noqa: F401
 
-__version__ = '5.0.0beta8'
+__version__ = '5.0.0beta9'
 
 
 def version():